when panning, block all mouse events on Canvas
When user is panning the canvas, all mouse events on the elements
shouldn't be triggered.
Extract two side-effect hooks from PipelineViewport
