+++ services/orchest-webserver/client/src/pipeline-view/pipeline-viewport/PipelineViewport.tsx
@@ -1,12 +1,11 @@
 import { Position } from "@/types";
 import { getHeight, getOffset, getWidth } from "@/utils/jquery-replacement";
 import { getScrollLineHeight } from "@/utils/webserver-utils";
-import { activeElementIsInput, uuidv4 } from "@orchest/lib-utils";
+import { uuidv4 } from "@orchest/lib-utils";
 import classNames from "classnames";
 import React from "react";
 import {
   DEFAULT_SCALE_FACTOR,
-  getScaleCorrectedPosition,
   originTransformScaling,
   scaleCorrected,
 } from "../common";
@@ -18,6 +17,8 @@ import { useValidateFilesOnSteps } from "../file-manager/useValidateFilesOnSteps
 import { INITIAL_PIPELINE_POSITION } from "../hooks/usePipelineCanvasState";
 import { STEP_HEIGHT, STEP_WIDTH } from "../PipelineStep";
 import { PipelineCanvas } from "./PipelineCanvas";
+import { useKeyboardEventsOnViewport } from "./useKeyboardEventsOnViewport";
+import { useMouseEventsOnViewport } from "./useMouseEventsOnViewport";
 
 const CANVAS_VIEW_MULTIPLE = 3;
 
@@ -47,7 +48,6 @@ const PipelineStepsOuterHolder: React.ForwardRefRenderFunction<
     mouseTracker,
     trackMouseMovement,
     dispatch,
-    keysDown,
     pipelineCwd,
     newConnection,
     environments,
@@ -134,47 +134,12 @@ const PipelineStepsOuterHolder: React.ForwardRefRenderFunction<
   );
 
   React.useEffect(() => {
-    const keyDownHandler = (event: KeyboardEvent) => {
-      if (activeElementIsInput()) return;
-
-      if (event.key === " " && !keysDown.has("Space")) {
-        // if any element is on focus, pressing space bar is equivalent to mouse click
-        // therefore it's needed to remove all "focus" state
-        if (document.activeElement instanceof HTMLElement) {
-          document.activeElement.blur();
-        }
-        setPipelineCanvasState({ panningState: "ready-to-pan" });
-        keysDown.add("Space");
-      }
-      if (event.key === "h" && !keysDown.has("h")) {
-        centerView();
-        keysDown.add("h");
-      }
-    };
-    const keyUpHandler = (event: KeyboardEvent) => {
-      if (event.key === " ") {
-        setPipelineCanvasState({ panningState: "idle" });
-        keysDown.delete("Space");
-      }
-      if (event.key === "h") {
-        keysDown.delete("h");
-      }
-    };
-
-    document.body.addEventListener("keydown", keyDownHandler);
-    document.body.addEventListener("keyup", keyUpHandler);
-    return () => {
-      document.body.removeEventListener("keydown", keyDownHandler);
-      document.body.removeEventListener("keyup", keyUpHandler);
-    };
-  }, [dispatch, keysDown, centerView, setPipelineCanvasState]);
-
-  React.useEffect(() => {
     if (
       pipelineOffset[0] === INITIAL_PIPELINE_POSITION[0] &&
       pipelineOffset[1] === INITIAL_PIPELINE_POSITION[1] &&
       eventVars.scaleFactor === DEFAULT_SCALE_FACTOR
     ) {
+      console.log("DEV WOW????");
       pipelineSetHolderOrigin([0, 0]);
     }
   }, [eventVars.scaleFactor, pipelineOffset, pipelineSetHolderOrigin]);
@@ -233,25 +198,36 @@ const PipelineStepsOuterHolder: React.ForwardRefRenderFunction<
   };
 
   const onMouseDown = (e: React.MouseEvent) => {
-    const isLeftClick = e.button === 0;
-
-    trackMouseMovement(e.clientX, e.clientY);
-
-    if (isLeftClick && panningState === "ready-to-pan") {
-      // space held while clicking, means canvas drag
-      setPipelineCanvasState({ panningState: "panning" });
+    if (eventVars.selectedConnection) {
+      dispatch({ type: "DESELECT_CONNECTION" });
     }
-
-    dispatch({ type: "DESELECT_CONNECTION" });
-
     // not dragging the canvas, so user must be creating a selection rectangle
     // we need to save the offset of cursor against pipeline canvas
-    if (isLeftClick && panningState === "idle") {
+    if (e.button === 0 && panningState === "idle") {
+      trackMouseMovement(e.clientX, e.clientY);
       dispatch({
         type: "CREATE_SELECTOR",
         payload: getOffset(canvasRef.current),
       });
     }
+  };
+
+  const onMouseUp = () => {
+    if (eventVars.stepSelector.active) {
+      dispatch({ type: "SET_STEP_SELECTOR_INACTIVE" });
+    } else {
+      dispatch({ type: "SELECT_STEPS", payload: { uuids: [] } });
+    }
+
+    if (eventVars.openedStep) {
+      dispatch({ type: "SET_OPENED_STEP", payload: undefined });
+    }
+
+    if (newConnection.current) {
+      dispatch({ type: "REMOVE_CONNECTION", payload: newConnection.current });
+    }
+
+    if (dragFile) onDropFiles();
   };
 
   const getApplicableStepFiles = useValidateFilesOnSteps();
@@ -299,92 +275,8 @@ const PipelineStepsOuterHolder: React.ForwardRefRenderFunction<
     createStepsWithFiles(dropPosition);
   }, [createStepsWithFiles, getOnCanvasPosition]);
 
-  const onMouseUp = (e: React.MouseEvent) => {
-    if (eventVars.stepSelector.active) {
-      dispatch({ type: "SET_STEP_SELECTOR_INACTIVE" });
-    } else {
-      dispatch({ type: "SELECT_STEPS", payload: { uuids: [] } });
-    }
-
-    if (eventVars.openedStep) {
-      dispatch({ type: "SET_OPENED_STEP", payload: undefined });
-    }
-
-    if (newConnection.current) {
-      dispatch({ type: "REMOVE_CONNECTION", payload: newConnection.current });
-    }
-
-    if (dragFile) onDropFiles();
-
-    const isLeftClick = e.button === 0;
-
-    if (isLeftClick && panningState === "panning") {
-      setPipelineCanvasState({ panningState: "ready-to-pan" });
-    }
-  };
-
-  const hasMouseMoved = React.useRef(false);
-  const onMouseMoveViewport = React.useCallback(() => {
-    if (!hasMouseMoved.current) {
-      // ensure that mouseTracker is in sync, to prevent jumping in some cases.
-      hasMouseMoved.current = true;
-      return;
-    }
-    let canvasOffset = getOffset(canvasRef.current);
-    // update newConnection's position
-    if (newConnection.current) {
-      const { x, y } = getScaleCorrectedPosition({
-        offset: canvasOffset,
-        position: mouseTracker.current.client,
-        scaleFactor: eventVars.scaleFactor,
-      });
-
-      newConnection.current = { ...newConnection.current, xEnd: x, yEnd: y };
-    }
-
-    if (eventVars.stepSelector.active) {
-      dispatch({ type: "UPDATE_STEP_SELECTOR", payload: canvasOffset });
-    }
-
-    if (panningState === "panning") {
-      let dx = mouseTracker.current.unscaledDelta.x;
-      let dy = mouseTracker.current.unscaledDelta.y;
-
-      setPipelineCanvasState((current) => ({
-        pipelineOffset: [
-          current.pipelineOffset[0] + dx,
-          current.pipelineOffset[1] + dy,
-        ],
-      }));
-    }
-  }, [
-    dispatch,
-    canvasRef,
-    eventVars.scaleFactor,
-    eventVars.stepSelector.active,
-    mouseTracker,
-    newConnection,
-    panningState,
-    setPipelineCanvasState,
-  ]);
-
-  const onMouseLeaveViewport = React.useCallback(() => {
-    if (eventVars.stepSelector.active) {
-      dispatch({ type: "SET_STEP_SELECTOR_INACTIVE" });
-    }
-    if (newConnection.current) {
-      dispatch({ type: "REMOVE_CONNECTION", payload: newConnection.current });
-    }
-  }, [dispatch, eventVars.stepSelector.active, newConnection]);
-
-  React.useEffect(() => {
-    document.body.addEventListener("mousemove", onMouseMoveViewport);
-    document.body.addEventListener("mouseleave", onMouseLeaveViewport);
-    return () => {
-      document.body.removeEventListener("mousemove", onMouseMoveViewport);
-      document.body.removeEventListener("mouseleave", onMouseLeaveViewport);
-    };
-  }, [onMouseLeaveViewport, onMouseMoveViewport]);
+  useMouseEventsOnViewport();
+  useKeyboardEventsOnViewport(canvasFuncRef);
 
   React.useEffect(() => {
     pipelineSetHolderSize();
